requirements.txt
PyOpenGL-3.1.5-cp38-cp38-win32.whl
PyOpenGL_accelerate-3.1.5-cp38-cp38-win32.whl